requirements.txt
psycopg2-binary
asterisk-ami

